Move version package to util/version This fixes a conflict between version/ and VERSION in case insensitive file systems
